requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
dnspython==2.1.0
idna==3.3
pexpect==4.8.0
ptyprocess==0.7.0
pycodestyle==2.8.0
requests==2.26.0
toml==0.10.2
urllib3==1.26.7
